gh-46  some refactoring
