--- Multisig/UI/Settings/BottomBorder.swift
@@ -1,9 +1,0 @@
-//
-//  BottomBorder.swift
-//  Multisig
-//
-//  Created by Dmitry Bespalov on 14.05.20.
-//  Copyright © 2020 Gnosis Ltd. All rights reserved.
-//
-
-import Foundation
